Create JaniscompLS470
Used for comparing the actual temperature of a sample mounted in the Janis (a LS470 series temp diode in this case taken in a 4wire measurement) and the Lakeshore Temperature Controller reported temperature. Utilizes RvTcsvcrunch. If having trouble with import, you can paste this code into a cell below and set the "crunch" function first.